batch-size 4

+++ src/masked_metrics.py
@@ -8,22 +8,22 @@ import keras
 
 
 def masked_accuracy(y_true, y_pred):
-    mask = tf.cast(tf.not_equal(y_true, -1), tf.float32)
-    y_pred_bin = tf.cast(y_pred > 0.5, tf.float32)
-    correct = tf.cast(tf.equal(y_true, y_pred_bin), tf.float32)
+    mask = tf.cast(tf.not_equal(y_true, -1), tf.float16)
+    y_pred_bin = tf.cast(y_pred > 0.5, tf.float16)
+    correct = tf.cast(tf.equal(y_true, y_pred_bin), tf.float16)
     return tf.reduce_sum(correct * mask) / tf.reduce_sum(mask + K.epsilon())
 
 def masked_precision(y_true, y_pred):
-    mask = tf.cast(tf.not_equal(y_true, -1), tf.float32)
-    y_pred_bin = tf.cast(y_pred > 0.5, tf.float32)
-    true_positives = tf.reduce_sum(tf.cast(y_pred_bin * y_true, tf.float32) * mask)
+    mask = tf.cast(tf.not_equal(y_true, -1), tf.float16)
+    y_pred_bin = tf.cast(y_pred > 0.5, tf.float16)
+    true_positives = tf.reduce_sum(tf.cast(y_pred_bin * y_true, tf.float16) * mask)
     predicted_positives = tf.reduce_sum(y_pred_bin * mask)
     return true_positives / (predicted_positives + K.epsilon())
 
 def masked_recall(y_true, y_pred):
-    mask = tf.cast(tf.not_equal(y_true, -1), tf.float32)
-    y_pred_bin = tf.cast(y_pred > 0.5, tf.float32)
-    true_positives = tf.reduce_sum(tf.cast(y_pred_bin * y_true, tf.float32) * mask)
+    mask = tf.cast(tf.not_equal(y_true, -1), tf.float16)
+    y_pred_bin = tf.cast(y_pred > 0.5, tf.float16)
+    true_positives = tf.reduce_sum(tf.cast(y_pred_bin * y_true, tf.float16) * mask)
     possible_positives = tf.reduce_sum(y_true * mask)
     return true_positives / (possible_positives + K.epsilon())
 
@@ -62,8 +62,8 @@ class MaskedAUC(tf.keras.metrics.AUC):
         return super(MaskedAUC, self).result()
 
 def masked_f1_score(y_true, y_pred):
-    mask = tf.cast(tf.not_equal(y_true, -1), tf.float32)
-    y_pred_bin = tf.cast(y_pred > 0.5, tf.float32)
+    mask = tf.cast(tf.not_equal(y_true, -1), tf.float16)
+    y_pred_bin = tf.cast(y_pred > 0.5, tf.float16)
 
     tp = tf.reduce_sum(y_pred_bin * y_true * mask)
     predicted_positives = tf.reduce_sum(y_pred_bin * mask)
